refactor(validators): extract repeated test into helper function
